Don't use the specification and version qualifiers to see if a package is needed
When installing the requirements of a package, we now just check that the name
of the requirement we are installing matches the name of one of the
dependencies of the package we are trying to satisfy the dependencies.

Previously, if we were trying to install the dependency "foo==1.0" of the
package "bar", which requires only "foo", buildout.dumppickedversions failed to
detect that "foo" was actually required by "bar" and the versions generated
didt have this information.

Additionnaly, we record the name of the package in lower case, so that it's easier
to retrieve an compare it.

git-svn-id: https://svn.plone.org/svn/collective/buildout/buildout.dumppickedversions/branches/multani-fix-generation-stability@236417 db7f04ef-aaf3-0310-a811-c281ed44c4ad